moved WritePort to myparallel
